@CurrentUser integrated, fixed overall issues

+++ todo/src/main/java/com/benchmark/todo/todo/TodoService.java
@@ -3,11 +3,11 @@ package com.benchmark.todo.todo;
 import com.benchmark.todo._core.error.CommonException;
 import com.benchmark.todo._core.error.ErrorCode;
 import com.benchmark.todo.todo.TodoDTO.Slim;
+import com.benchmark.todo.user.entity.User;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,19 +18,25 @@ public class TodoService {
 
   private final TodoRepository todoRepository;
 
-  // TODO: delete this code
-  public Todo getTodoById(long id) {
-    // TODO: fix the exception
-    return todoRepository.findById(id).orElseThrow(IllegalStateException::new);
+  public List<Todo> fetchAll() {
+    return todoRepository.findAll();
   }
 
-  public TodoDTO.Slim getTodoSlimById(long id) {
-    Todo todo = getTodoById(id);
+  public TodoDTO.Slim fetchTodoSlimById(User user, long id) {
+
+    Todo todo = todoRepository.findById(id).orElseThrow(() ->
+        new CommonException(ErrorCode.NOT_FOUND));
+
+    // Is this right way of checking ownership?
+    if (!todo.isOwnedBy(user.getId())) {
+      throw new CommonException(ErrorCode.UNAUTHORIZED);
+    }
+
     return TodoDTO.Slim.of(todo);
   }
 
-  public List<TodoDTO.Slim> getAllTodoSlims() {
-    List<Todo> todoList = todoRepository.findAll();
+  public List<TodoDTO.Slim> fetchAllTodoSlims(User user) {
+    List<Todo> todoList = todoRepository.findAllByUserId(user.getId());
 
     return todoList
         .stream()
@@ -38,23 +44,26 @@ public class TodoService {
         .toList();
   }
 
-  public void createTodo(TodoDTO.CreateRequest createDto) {
-    todoRepository.save(createDto.toEntity());
+  public void createTodo(User user, TodoDTO.CreateRequest createDto) {
+    todoRepository.save(createDto.toEntity(user.getId()));
   }
 
-  public void updateTodo(long id, TodoDTO.UpdateRequest updateDto) {
+  public void updateTodo(User user, long id, TodoDTO.UpdateRequest updateDto) {
 
     Todo todo = todoRepository.findById(id).orElseThrow(() ->
         new CommonException(ErrorCode.NOT_FOUND));
 
-//    Todo todo = getTodoById(id);
+    if (!todo.isOwnedBy(user.getId())) {
+      throw new CommonException(ErrorCode.UNAUTHORIZED);
+    }
+
     todo.setDueDate(updateDto.getDueDate());
     todo.setDescription(updateDto.getDescription());
     todoRepository.save(todo);
   }
 
   // TODO: use custom query
-  public List<LocalDate> getCalenderDates(int month) {
+  public List<LocalDate> fetchCalenderDates(int month) {
     Set<LocalDate> dates = new HashSet<>();
     List<Todo> todoList = new ArrayList<>(
         todoRepository.findAll()
@@ -71,12 +80,18 @@ public class TodoService {
     return dateList;
   }
 
-  public void deleteTodo(long id) {
+  public void deleteTodo(User user, long id) {
+    Todo todo = todoRepository.findById(id).orElseThrow(() ->
+        new CommonException(ErrorCode.NOT_FOUND));
+
+    if (!todo.isOwnedBy(user.getId())) {
+      throw new CommonException(ErrorCode.UNAUTHORIZED);
+    }
     todoRepository.deleteById(id);
   }
 
-  public List<TodoDTO.Slim> findTodoByDate(LocalDate date) {
-    List<Todo> todos = todoRepository.findAll();
+  public List<TodoDTO.Slim> fetchTodoByDate(User user, LocalDate date) {
+    List<Todo> todos = todoRepository.findAllByUserId(user.getId());
 
     return todos.stream()
         .filter(todo -> todo.getDueDate().toLocalDate().equals(date))
